First pass at a quirks framework.
Now we actually send fully qualified HTTP/S URLs back when a DAV:href
is needed -- except for the MacOS 10.6 case.

Quirks are defined in config/quirks.yml and shoved into a hash in the
Quirks class by the config/quirks.rb initializer.  The [] operator is
used on the Quirks class to retrieve individual quirks.  For now,
quirks are simply an array of regular expressions that we can attempt to
match against the user agent string.

Signed-off-by: Alex Zepeda <alex@inferiorhumanorgans.com>